Minor fix in Rakefile
worldstate_loader: no longer being bundled by `rake all`. Use
worldstate_picker instead

+++ Rakefile
@@ -81,7 +81,7 @@ task :cleanup do
     puts "Removing old documentation: #{bundledFile}"
     File.delete bundledFile
   end
-  File.delete 'documentation.htm'
+  File.delete 'documentation.htm' if File.exists? 'documentation.htm'
 end
 
 desc 'Create all zipped Wirecloud widget files'
